refactored places funded for to decimal

+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Models/SetSchoolOverviewModel.cs
@@ -6,7 +6,7 @@
       public string PublishedAdmissionNumber { get; set; }
       public string ViabilityIssues { get; set; }
       public string FinancialDeficit { get; set; }
-      public int? NumberOfPlacesFundedFor { get; set; }
+      public decimal? NumberOfPlacesFundedFor { get; set; }
       public string PartOfPfiScheme { get; set; }
       public string PfiSchemeDetails { get; set; }
       public decimal? DistanceFromSchoolToTrustHeadquarters { get; set; }
@@ -20,7 +20,7 @@
           string publishedAdmissionNumber,
           string viabilityIssues,
           string financialDeficit,
-          int? numberOfPlacesFundedFor,
+          decimal? numberOfPlacesFundedFor,
           string partOfPfiScheme,
           string pfiSchemeDetails,
           decimal? distanceFromSchoolToTrustHeadquarters,
